0.19.5 new package for Height + refactoring
