+++ wfrp/src/main/java/com/warhammer/wfrp/height/HeightDto.java
@@ -1,4 +1,4 @@
-package com.warhammer.wfrp.dto.appearance;
+package com.warhammer.wfrp.height;
 
 public record HeightDto(String height) {
